[stinkytofu] Fix sign-compare warnings in WaitCntInsertionPass
Cast int wait counts to size_t before comparing with deque::size().
The values are guaranteed non-negative at the comparison point (guarded
by checks for kUnused and 0).

+++ shared/stinkytofu/src/transforms/asm/StinkyWaitCntInsertionPass.cpp
@@ -387,13 +387,13 @@ class StinkyWaitCntInsertionPass : public StinkyInstPass {
         }
 
         if (lastDSWaitCount != WaitCntInstruction::kUnused && lastDSWaitCount != 0 &&
-            lastDSWaitCount < currentBlockState.dsQueue.size()) {
+            static_cast<size_t>(lastDSWaitCount) < currentBlockState.dsQueue.size()) {
             currentBlockState.dsQueue.erase(currentBlockState.dsQueue.begin(),
                                             currentBlockState.dsQueue.end() - lastDSWaitCount);
         }
 
         if (lastBufferWaitCount != WaitCntInstruction::kUnused && lastBufferWaitCount != 0 &&
-            lastBufferWaitCount < currentBlockState.bufferQueue.size()) {
+            static_cast<size_t>(lastBufferWaitCount) < currentBlockState.bufferQueue.size()) {
             currentBlockState.bufferQueue.erase(
                 currentBlockState.bufferQueue.begin(),
                 currentBlockState.bufferQueue.end() - lastBufferWaitCount);
